refactor(MsTeamsConnectButton & SlackConnectButton): streamline integrationIdentifier handling and remove unused constants
- Updated the MsTeamsConnectButton and SlackConnectButton components to require integrationIdentifier as a mandatory prop, enhancing type safety.
- Removed deprecated constants related to MS Teams and Slack integrations, simplifying the codebase.
- Adjusted connection identifier handling to improve clarity and maintainability across components.
