+++ packages/js/src/ui/components/constants.ts
@@ -1,2 +1,2 @@
 export const DEFAULT_MSTEAMS_CONNECTION_IDENTIFIER = 'chconn-msteams-default';
-export const DEFAULT_MSTEAMS_INTEGRATION_IDENTIFIER = 'msteams';
+export const DEFAULT_SLACK_CONNECTION_IDENTIFIER = 'chconn-slack-default';
