--- packages/js/src/ui/components/slack-constants.ts
@@ -1,2 +1,0 @@
-export const DEFAULT_CONNECTION_IDENTIFIER = 'chconn-slack-default';
-export const DEFAULT_INTEGRATION_IDENTIFIER = 'slack';
